backend/batch/requirements.txt
google-api-python-client>=2.100.0
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0
boto3>=1.28.0
